Modify ED,VCA and FreeSystem to fit the new protocals of engine.

+++ Math/test/Tensor.py
@@ -9,7 +9,7 @@ from HamiltonianPy.Math.Tensor import *
 
 def test_tensor():
     print 'test_tensor'
-    b=Tensor(random.random((2,2,3)),labels=['i','j','k'],dtype=float64)
+    b=Tensor(random.random((2,2,3)),labels=['i','j','k'])
     print "b: %s"%b
     print "b.transpose(axes=[2,1,0]): %s"%b.transpose(axes=[2,1,0])
     print "b.transpose(labels=['i','j','k']): %s"%b.transpose(labels=['i','j','k'])
